Refactor compiler errors and warnings handling - Part 1 (#61)
This patch is the initial part of the effort of consolidating the handling of compiler warnings and errors across compiler components. The changes in this patch are meant to be gentle suboptimal in order to restrict the change relatively small and easier to digest, test, and review. Subsequently changes on refactoring the process with better practices should be followed.