Refine integration test compiler error handling and checking (#64)
This patch follows the work done in #62 and #63 by refining the way compiler errors are archived and handled in the integration tests. This change makes it a little easier to maintain the associated expected compiler errors of each integration test case by storing the expected compiler error data in a more structured format.